+++ src/StardewModdingAPI/LogWriter.cs
@@ -60,7 +60,7 @@ namespace StardewModdingAPI
         /// <summary>Raise a deprecation warning.</summary>
         private void WarnDeprecated()
         {
-            Program.DeprecationManager.Warn($"the {nameof(LogWriter)} class", "1.0", DeprecationLevel.Info);
+            Program.DeprecationManager.Warn($"the {nameof(LogWriter)} class", "1.0", DeprecationLevel.PendingRemoval);
         }
     }
 }